# docs/specification/openapi.yaml
openapi: 3.1.0
info:
  title: TestModeller API
  version: 0.1.0
  description: >
    Contract between the Angular UI and the Rust backend. Changes require explicit approval (see AGENTS.md).
    Hierarchy: Project > Component > Feature > Model. A Feature has a scenario description and one or more Models
    (state machines). Test cases belong to a Feature and are assigned to Model states or transitions (steps).
servers:
  - url: /api/v1
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List projects
      parameters: [ { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Cursor' } ]
      responses:
        '200': { description: Page of projects, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectPage' } } } }
    post:
      tags: [Projects]
      summary: Create project
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '400': { $ref: '#/components/responses/Problem' }
  /projects/{projectId}:
    parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
    get:
      tags: [Projects]
      summary: Get project
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Projects]
      summary: Update project
      parameters: [ { $ref: '#/components/parameters/IfMatch' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectInput' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Projects]
      summary: Delete project and all its contents
      responses:
        '204': { description: Deleted }
  /projects/{projectId}/tree:
    parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
    get:
      tags: [Projects]
      summary: Component > Feature > Model tree for the explorer
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectTree' } } } }
...
  /projects/{projectId}/components:
    parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
    get:
      tags: [Components]
      summary: List components
      parameters: [ { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Cursor' } ]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ComponentPage' } } } }
    post:
      tags: [Components]
      summary: Create component
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ComponentInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Component' } } } }
  /components/{componentId}:
    parameters: [ { $ref: '#/components/parameters/ComponentId' } ]
    get:
      tags: [Components]
      summary: Get component
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Component' } } } }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Components]
      summary: Update component
      parameters: [ { $ref: '#/components/parameters/IfMatch' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ComponentInput' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Component' } } } }
        '412': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Components]
      summary: Delete component (409 if it has features unless cascade=true)
      parameters: [ { $ref: '#/components/parameters/Cascade' } ]
      responses:
        '204': { description: Deleted }
        '409': { $ref: '#/components/responses/Problem' }
  /components/{componentId}/features:
    parameters: [ { $ref: '#/components/parameters/ComponentId' } ]
    get:
      tags: [Features]
      summary: List features of a component
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
        - { name: tag, in: query, schema: { type: string } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/FeaturePage' } } } }
    post:
      tags: [Features]
      summary: Create feature
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/FeatureInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Feature' } } } }
  /features/{featureId}:
    parameters: [ { $ref: '#/components/parameters/FeatureId' } ]
    get:
      tags: [Features]
      summary: Get feature including its scenario description
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Feature' } } } }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Features]
      summary: Update feature (name, description, scenarioDescription, tags)
      parameters: [ { $ref: '#/components/parameters/IfMatch' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/FeatureInput' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Feature' } } } }
        '412': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Features]
      summary: Delete feature (409 if it has models or test cases unless cascade=true)
      parameters: [ { $ref: '#/components/parameters/Cascade' } ]
      responses:
        '204': { description: Deleted }
        '409': { $ref: '#/components/responses/Problem' }
  /features/{featureId}/move:
    parameters: [ { $ref: '#/components/parameters/FeatureId' } ]
    post:
      tags: [Features]
      summary: Move feature to another component
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [componentId], properties: { componentId: { $ref: '#/components/schemas/Id' } } } } }
      responses:
        '200': { description: Moved, content: { application/json: { schema: { $ref: '#/components/schemas/Feature' } } } }
...
components:
  parameters:
    ProjectId: { name: projectId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    ComponentId: { name: componentId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
    FeatureId: { name: featureId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Entity `version` for optimistic concurrency
      schema: { type: integer, minimum: 1 }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
    Cursor: { name: cursor, in: query, schema: { type: string } }
    Cascade: { name: cascade, in: query, schema: { type: boolean, default: false } }
  responses:
    Problem:
      description: Error (RFC 7807)
      content:
        application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
  schemas:
    Id: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time }
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        errors:
          type: array
          items: { type: object, properties: { field: { type: string }, message: { type: string } } }
...
    Audit:
      type: object
      required: [id, version, createdAt, updatedAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        version: { type: integer, minimum: 1 }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
    ProjectInput:
      type: object
      required: [name]
      properties: { name: { type: string, minLength: 1, maxLength: 200 }, description: { type: string } }
    Project:
      allOf: [ { $ref: '#/components/schemas/Audit' }, { $ref: '#/components/schemas/ProjectInput' } ]
    ProjectPage:
      type: object
      properties: { items: { type: array, items: { $ref: '#/components/schemas/Project' } }, nextCursor: { type: [string, 'null'] } }
    ComponentInput:
      type: object
      required: [name]
      properties: { name: { type: string, minLength: 1, maxLength: 200 }, description: { type: string } }
    Component:
      allOf:
        - { $ref: '#/components/schemas/Audit' }
        - { $ref: '#/components/schemas/ComponentInput' }
        - { type: object, required: [projectId], properties: { projectId: { $ref: '#/components/schemas/Id' } } }
    ComponentPage:
      type: object
      properties: { items: { type: array, items: { $ref: '#/components/schemas/Component' } }, nextCursor: { type: [string, 'null'] } }
    FeatureInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string }
        scenarioDescription: { type: string, description: Free-text (markdown) description of the scenarios this feature's models represent }
        tags: { type: array, items: { type: string } }
    Feature:
      allOf:
        - { $ref: '#/components/schemas/Audit' }
        - { $ref: '#/components/schemas/FeatureInput' }
        - { type: object, required: [componentId], properties: { componentId: { $ref: '#/components/schemas/Id' } } }
    FeaturePage:
      type: object
      properties: { items: { type: array, items: { $ref: '#/components/schemas/Feature' } }, nextCursor: { type: [string, 'null'] } }
    ModelStatus: { type: string, enum: [draft, ready] }
...
    ProjectTree:
      type: object
      properties:
        components:
          type: array
          items:
            type: object
            properties:
              id: { $ref: '#/components/schemas/Id' }
              name: { type: string }
              features:
                type: array
                items:
                  type: object
                  properties:
                    id: { $ref: '#/components/schemas/Id' }
                    name: { type: string }
                    models: { type: array, items: { type: object, properties: { id: { $ref: '#/components/schemas/Id' }, name: { type: string }, status: { $ref: '#/components/schemas/ModelStatus' } } } }
...